requirements.txt
torch
torchvision
torchmetrics
wandb
numpy
matplotlib
Pillow
scikit-learn
tqdm
omegaconf
